+++ frontend/src/pages/Graphing/Combobox/index.tsx
@@ -22,6 +22,7 @@ type Props = {
 	}
 	onlyNumericKeys?: boolean
 	defaultKeys?: string[]
+	disabled?: boolean
 }
 
 export const Combobox: React.FC<Props> = ({
@@ -31,6 +32,7 @@ export const Combobox: React.FC<Props> = ({
 	searchConfig: { productType, startDate, endDate },
 	onlyNumericKeys,
 	defaultKeys,
+	disabled,
 }) => {
 	const { projectId } = useProjectId()
 	const [getKeys, { data }] = useGetKeysLazyQuery()
@@ -95,13 +97,16 @@ export const Combobox: React.FC<Props> = ({
 		<ComboboxSelect
 			label={label}
 			value={selection}
-			valueRender={<Text cssClass={style.comboboxText}>{selection}</Text>}
+			valueRender={
+				<Text cssClass={style.comboboxText}>{selection || 'Rows'}</Text>
+			}
 			options={keyOptions}
 			onChange={setSelection}
 			onChangeQuery={setQuery}
 			cssClass={style.combobox}
 			wrapperCssClass={style.comboboxWrapper}
 			queryPlaceholder="Filter..."
+			disabled={disabled}
 		/>
 	)
 }