+++ frontend/src/pages/Player/TimelineIndicatorsContext/TimelineIndicatorsContext.tsx
@@ -2,7 +2,6 @@ import { useGetTimelineIndicatorEventsQuery } from '@graph/hooks'
 import { Session, TimelineIndicatorEvent } from '@graph/schemas'
 import { LoadingError } from '@pages/Player/ResourcesContext/ResourcesContext'
 import { indexedDBFetch } from '@util/db'
-import { checkResourceLimit } from '@util/preload'
 import { H } from 'highlight.run'
 import { useEffect, useState } from 'react'
 
@@ -56,18 +55,6 @@ export const useTimelineIndicators = (
 			setLoading(true)
 			;(async () => {
 				if (!session.timeline_indicators_url) return
-				const limit = await checkResourceLimit(
-					session.timeline_indicators_url,
-				)
-				if (limit) {
-					setError(LoadingError.NetworkResourcesTooLarge)
-					H.consumeError(new Error(limit.error), undefined, {
-						fileSize: limit.fileSize.toString(),
-						limit: limit.sizeLimit.toString(),
-						sessionSecureID: session?.secure_id,
-					})
-					return
-				}
 				let response
 				for await (const r of indexedDBFetch(
 					session.timeline_indicators_url,