+++ sdk/highlight-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @highlight-run/react
 
+## 6.0.1
+
+### Patch Changes
+
+-   Updated dependencies [a95d52b]
+-   Updated dependencies [f432e66]
+    -   highlight.run@9.4.1
+
+## 6.0.0
+
+### Patch Changes
+
+-   Updated dependencies [815faa8]
+-   Updated dependencies [815faa8]
+    -   highlight.run@9.4.0
+
 ## 5.0.4
 
 ### Patch Changes